feat : 지점 상세 목록 출력
